pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sermon-summarizer"
version = "0.1.0"
description = "Automated pipeline to download, process, transcribe, and format church sermons from YouTube into structured markdown"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Sermon Summarizer Team", email = "team@example.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.1.0",
    "pydub>=0.25.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "click>=8.1.0",
    "pathlib2>=2.3.0",
    "PyYAML>=6.0",
    "openai>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "factory-boy>=3.3.0",
    "hypothesis>=6.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "factory-boy>=3.3.0",
    "hypothesis>=6.0.0",
    "PyYAML>=6.0",
]

[project.scripts]
sermon-summarizer = "sermon_summarizer.cli:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/sermon-summarizer"
"Bug Tracker" = "https://github.com/yourusername/sermon-summarizer/issues"


[tool.setuptools]
packages = ["sermon_summarizer"]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.package-data]
sermon_summarizer = ["*.yaml", "*.yml", "*.json", "py.typed"]

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | \.pytest_cache
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503", "E501"]
exclude = [
    ".git",
    "__pycache__",
    ".pytest_cache",
    "build",
    "dist",
    "*.egg-info",
    "venv",
    ".venv"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=sermon_summarizer",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=60"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests requiring cluster access",
    "unit: marks tests as unit tests (default)",
]

[tool.coverage.run]
source = ["src/sermon_summarizer"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "yt_dlp.*",
    "pydub.*",
    "factory.*",
    "hypothesis.*",
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
